readme: replace headings with alerts
Used GitHub Flavored Markdown's alerts feature for more semantic markup than using headings. For all the reasons semantic markup is usually good, but on GitHub in particular they add underlines and links and stuff to headings which aren't needed here.

Arguably makes the text stand out less than using headings, but when when everything is a heading, it's hard for anything to stand out.

Used bold and different kinds of alerts to indicate relative importance.

Otherwise, tried to keep refactoring to a minimum.